+++ src/index.ts
@@ -1,2 +1,3 @@
 export * from './array/random';
 export * from './array/shuffle';
+export * from "./array/pickN";